Test to see the need for monitoring
Althoght now the GenServer is monitorized handle_cast/2 implementation, you can see the registry is both linking and monitoring the buckets:
{:ok, pid} = KV.Bucket.start_link([])
ref = Process.monitor(pid)

This is a bad idea, as we don’t want the registry to crash when a bucket crashes! We typically avoid creating new processes directly, instead, we delegate this responsibility to supervisors.